refactor: extract filterValidSymlinks to utils
The symlink validation logic (realpath containment, stat().isFile(),
broken-symlink handling) was duplicated between search-index.ts and
vault-filesystem.ts. Extracted to a shared utils helper that takes a
warn callback, keeping it generic with zero domain knowledge.
