+++ src/utils/filter-valid-symlinks.ts
@@ -1,0 +1,47 @@
+/** Filters directory entries, excluding symlinks whose resolved targets are
+ *  broken (dangling), escape the given root directory, or are not regular files. */
+
+import type { Dirent } from "node:fs"
+import { realpath, stat } from "node:fs/promises"
+import { join, relative } from "node:path"
+import { describeError } from "./describe-error.js"
+
+type WarnFn = (message: string, meta: Record<string, unknown>) => void
+
+export const filterValidSymlinks = async (
+  entries: ReadonlyArray<Dirent>,
+  canonicalRoot: string,
+  normalizedRoot: string,
+  warn: WarnFn,
+): Promise<Dirent[]> =>
+  (
+    await Promise.all(
+      entries.map(async (entry) => {
+        if (!entry.isSymbolicLink()) return entry
+        const entryPath = join(entry.parentPath, entry.name)
+        try {
+          const targetPath = await realpath(entryPath)
+          if (!targetPath.startsWith(canonicalRoot + "/")) {
+            warn("symlink target escapes root, skipping", {
+              path: relative(normalizedRoot, entryPath),
+            })
+            return null
+          }
+          const targetStat = await stat(targetPath)
+          if (!targetStat.isFile()) {
+            warn("symlink target is not a file, skipping", {
+              path: relative(normalizedRoot, entryPath),
+            })
+            return null
+          }
+          return entry
+        } catch (error) {
+          warn("broken symlink, skipping", {
+            path: relative(normalizedRoot, entryPath),
+            error: describeError(error),
+          })
+          return null
+        }
+      }),
+    )
+  ).filter((entry): entry is Dirent => entry !== null)
